Refactored actions into three separate classes
